+++ cider-repl.el
@@ -227,6 +227,8 @@ Insert a banner, unless NOPROMPT is non-nil."
     ;; use the same requires by default as clojure.main does
     (cider-eval-sync nrepl-repl-requires-sexp)
     (cider-repl-reset-markers)
+    ;; honor :init-ns from lein's :repl-options on startup
+    (setq nrepl-buffer-ns (cider-eval-and-get-value "(str *ns*)"))
     (unless noprompt
       (cider-repl--insert-banner-and-prompt nrepl-buffer-ns))
     (cider-remember-clojure-buffer cider-current-clojure-buffer)